LayoutTests/fast/events/ios/autocorrect-with-caret-selection.html

<!DOCTYPE html> <!-- webkit-test-runner [ useFlexibleViewport=true ] -->
<html>
<head>
<meta name="viewport" content="initial-scale=1.0, user-scalable=no">
<script src="../../../resources/ui-helper.js"></script>
<script src="../../../resources/js-test.js"></script>
<style>
#editable {
    border: 1px solid tomato;
    display: inline-block;
    font-size: 32px;
}
</style>
<script>
jsTestIsAsync = true;
addEventListener("load", async () => {
    description("To manually test, focus the editable element above, move the selection to the end, and select any text suggestion that appears in the keyboard. The text suggestion should be inserted.");

    editable = document.getElementById("editable");
    await UIHelper.setHardwareKeyboardAttached(false);
    await UIHelper.activateElementAndWaitForInputSession(editable);

    getSelection().setPosition(editable, 1);
    await UIHelper.applyAutocorrection("world", "");

    shouldBeEqualToString("editable.textContent", "Hello world");
    finishJSTest();
});
</script>
</head>
<body>
<div contenteditable id="editable">Hello&nbsp;</div>
</body>
</html>
